backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
pydantic
langchain
langchain-google-genai
firebase-admin
google-cloud-firestore